refactor(profile): split health into directory module
Splits crates/crosshook-core/src/profile/health.rs (1309 lines) into
a directory-backed module with per-responsibility submodules and a
tests/ subdirectory.

No behavior change — pure structural split. Public API unchanged:
HealthStatus, HealthIssueSeverity, HealthIssue, ProfileHealthReport,
HealthCheckSummary plus check_profile_health,
batch_check_health_with_enrich, batch_check_health,
build_dependency_health_issues remain exported under the same paths.

Part of #290 (task 16).
