--- src/crosshook-native/crates/crosshook-core/src/profile/health.rs
@@ -1,1309 +1,0 @@
-use std::fs;
-use std::io::ErrorKind;
-
-use chrono::Utc;
-use serde::{Deserialize, Serialize};
-
-use crate::launch::request::{METHOD_PROTON_RUN, METHOD_STEAM_APPLAUNCH};
-use crate::metadata::PrefixDependencyStateRow;
-use crate::platform::{
-    normalize_flatpak_host_path, normalized_path_exists_on_host, normalized_path_is_dir,
-    normalized_path_is_dir_on_host, normalized_path_is_executable_file,
-    normalized_path_is_executable_file_on_host, normalized_path_is_file_on_host,
-};
-use crate::profile::models::{resolve_launch_method, GameProfile};
-use crate::profile::toml_store::ProfileStore;
-
-/// Profile-level health roll-up.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum HealthStatus {
-    Healthy,
-    Stale,
-    Broken,
-}
-
-/// Per-issue severity — distinct from `ValidationSeverity` which always returns Fatal.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum HealthIssueSeverity {
-    Error,
-    Warning,
-    Info,
-}
-
-/// A single path-field issue found during health check.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HealthIssue {
-    pub field: String,
-    pub path: String,
-    pub message: String,
-    pub remediation: String,
-    pub severity: HealthIssueSeverity,
-}
-
-/// Per-profile health check result.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProfileHealthReport {
-    pub name: String,
-    pub status: HealthStatus,
-    pub launch_method: String,
-    pub issues: Vec<HealthIssue>,
-    pub checked_at: String,
-}
-
-/// Batch health check result across all profiles.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HealthCheckSummary {
-    pub profiles: Vec<ProfileHealthReport>,
-    pub healthy_count: usize,
-    pub stale_count: usize,
-    pub broken_count: usize,
-    pub total_count: usize,
-    pub validated_at: String,
-}
-
-fn normalized_host_probe_path(raw_path: &str) -> String {
-    normalize_flatpak_host_path(raw_path).trim().to_string()
-}
-
-fn display_path(raw_path: &str) -> String {
-    let original = raw_path.trim();
-    if !original.is_empty() {
-        return original.to_string();
-    }
-    normalized_host_probe_path(raw_path)
-}
-
-fn path_exists_visible_or_host(raw_path: &str) -> bool {
-    let original = raw_path.trim();
-    if original.is_empty() {
-        return false;
-    }
-
-    if std::path::Path::new(original).exists() {
-        return true;
-    }
-
-    let normalized = normalized_host_probe_path(raw_path);
-    !normalized.is_empty()
-        && (std::path::Path::new(&normalized).exists()
-            || normalized_path_exists_on_host(&normalized))
-}
-
-fn path_is_file_visible_or_host(raw_path: &str) -> bool {
-    let original = raw_path.trim();
-    if original.is_empty() {
-        return false;
-    }
-
-    if std::path::Path::new(original).is_file() {
-        return true;
-    }
-
-    let normalized = normalized_host_probe_path(raw_path);
-    !normalized.is_empty()
-        && (std::path::Path::new(&normalized).is_file()
-            || normalized_path_is_file_on_host(&normalized))
-}
-
-fn path_is_dir_visible_or_host(raw_path: &str) -> bool {
-    let original = raw_path.trim();
-    if original.is_empty() {
-        return false;
-    }
-
-    if std::path::Path::new(original).is_dir() {
-        return true;
-    }
-
-    let normalized = normalized_host_probe_path(raw_path);
-    !normalized.is_empty()
-        && (normalized_path_is_dir(normalized.as_str())
-            || normalized_path_is_dir_on_host(normalized.as_str()))
-}
-
-fn path_is_executable_file(path: &str) -> bool {
-    let Ok(metadata) = fs::metadata(path) else {
-        return false;
-    };
-    if !metadata.is_file() {
-        return false;
-    }
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        metadata.permissions().mode() & 0o111 != 0
-    }
-
-    #[cfg(not(unix))]
-    {
-        true
-    }
-}
-
-fn path_is_executable_visible_or_host(raw_path: &str) -> bool {
-    let original = raw_path.trim();
-    if original.is_empty() {
-        return false;
-    }
-
-    if path_is_executable_file(original) {
-        return true;
-    }
-
-    let normalized = normalized_host_probe_path(raw_path);
-    !normalized.is_empty()
-        && (normalized_path_is_executable_file(normalized.as_str())
-            || normalized_path_is_executable_file_on_host(normalized.as_str()))
-}
-
-/// Classify a path check result (missing file vs. wrong type / inaccessible) into a `HealthIssue`.
-///
-/// Returns `None` when the path is healthy (present, correct type, accessible).
-fn check_file_path(
-    field: &str,
-    path: &str,
-    severity_on_broken: HealthIssueSeverity,
-) -> Option<(HealthIssue, bool /* is_stale */)> {
-    let display_path = display_path(path);
-    if display_path.is_empty() {
-        return None;
-    }
-
-    let original = path.trim();
-    if let Ok(meta) = fs::metadata(original) {
-        if meta.is_file() {
-            return None;
-        }
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("Path exists but is not a file: {display_path}"),
-                remediation: "Select the file itself, not a directory or other path type."
-                    .to_string(),
-                severity: severity_on_broken,
-            },
-            false,
-        ));
-    }
-
-    if path_is_file_visible_or_host(path) {
-        return None;
-    }
-
-    if path_exists_visible_or_host(path) {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("Path exists but is not a file: {display_path}"),
-                remediation: "Select the file itself, not a directory or other path type."
-                    .to_string(),
-                severity: severity_on_broken,
-            },
-            false,
-        ));
-    }
-
-    match fs::metadata(original) {
-        Ok(meta) if meta.is_file() => {
-            // Exists and is a file — healthy
-            None
-        }
-        Ok(_) => {
-            // Exists but wrong type (directory, symlink to dir, etc.)
-            Some((
-                HealthIssue {
-                    field: field.to_string(),
-                    path: display_path.to_string(),
-                    message: format!("Path exists but is not a file: {display_path}"),
-                    remediation: "Select the file itself, not a directory or other path type."
-                        .to_string(),
-                    severity: severity_on_broken,
-                },
-                false,
-            ))
-        }
-        Err(err) if err.kind() == ErrorKind::NotFound => {
-            // Missing from disk → Stale
-            Some((
-                HealthIssue {
-                    field: field.to_string(),
-                    path: display_path.to_string(),
-                    message: format!("Path does not exist: {display_path}"),
-                    remediation: "Re-browse to the file or verify the path is correct.".to_string(),
-                    severity: HealthIssueSeverity::Warning,
-                },
-                true,
-            ))
-        }
-        Err(err) if err.kind() == ErrorKind::PermissionDenied => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Path is not accessible (permission denied): {display_path}"),
-                remediation: "Check file permissions (e.g. chmod a+r).".to_string(),
-                severity: severity_on_broken,
-            },
-            false,
-        )),
-        Err(err) => {
-            // Other I/O errors are treated as broken
-            Some((
-                HealthIssue {
-                    field: field.to_string(),
-                    path: display_path.to_string(),
-                    message: format!("Could not access path: {err}"),
-                    remediation: "Verify the path is valid and accessible.".to_string(),
-                    severity: severity_on_broken,
-                },
-                false,
-            ))
-        }
-    }
-}
-
-/// Check a required file field. Empty path → `Broken`. Missing → `Stale`. Wrong type → `Broken`.
-fn check_required_file(field: &str, path: &str) -> Option<(HealthIssue, bool)> {
-    if path.trim().is_empty() {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: String::new(),
-                message: format!("Required field '{field}' is not configured."),
-                remediation: format!("Browse to or enter a value for '{field}'."),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-    check_file_path(field, path, HealthIssueSeverity::Error)
-}
-
-/// Check a required directory field. Empty → `Broken`. Missing → `Stale`. Wrong type → `Broken`.
-fn check_required_directory(field: &str, path: &str) -> Option<(HealthIssue, bool)> {
-    let display_path = display_path(path);
-    if display_path.is_empty() {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: String::new(),
-                message: format!("Required field '{field}' is not configured."),
-                remediation: format!("Browse to or enter a value for '{field}'."),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    let original = path.trim();
-    if let Ok(meta) = fs::metadata(original) {
-        if meta.is_dir() {
-            return None;
-        }
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("Path exists but is not a directory: {display_path}"),
-                remediation: "Select the directory itself, not a file inside it.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    if path_is_dir_visible_or_host(path) {
-        return None;
-    }
-
-    if path_exists_visible_or_host(path) {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("Path exists but is not a directory: {display_path}"),
-                remediation: "Select the directory itself, not a file inside it.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    match fs::metadata(original) {
-        Ok(meta) if meta.is_dir() => None,
-        Ok(_) => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Path exists but is not a directory: {display_path}"),
-                remediation: "Select the directory itself, not a file inside it.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        )),
-        Err(err) if err.kind() == ErrorKind::NotFound => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Directory does not exist: {display_path}"),
-                remediation: "Re-browse to the directory or verify the path is correct."
-                    .to_string(),
-                severity: HealthIssueSeverity::Warning,
-            },
-            true,
-        )),
-        Err(err) if err.kind() == ErrorKind::PermissionDenied => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Directory is not accessible (permission denied): {display_path}"),
-                remediation: "Check directory permissions (e.g. chmod a+rx).".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        )),
-        Err(err) => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Could not access directory: {err}"),
-                remediation: "Verify the path is valid and accessible.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        )),
-    }
-}
-
-/// Check a required executable file field. Empty → `Broken`. Missing → `Stale`. Not executable → `Broken`.
-fn check_required_executable(field: &str, path: &str) -> Option<(HealthIssue, bool)> {
-    let display_path = display_path(path);
-    if display_path.is_empty() {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: String::new(),
-                message: format!("Required field '{field}' is not configured."),
-                remediation: format!("Browse to or enter a value for '{field}'."),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    let original = path.trim();
-    if let Ok(meta) = fs::metadata(original) {
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            if !meta.is_file() {
-                return Some((
-                    HealthIssue {
-                        field: field.to_string(),
-                        path: display_path.clone(),
-                        message: format!("Path exists but is not a file: {display_path}"),
-                        remediation: "Select the executable file itself.".to_string(),
-                        severity: HealthIssueSeverity::Error,
-                    },
-                    false,
-                ));
-            }
-            if meta.permissions().mode() & 0o111 == 0 {
-                return Some((
-                    HealthIssue {
-                        field: field.to_string(),
-                        path: display_path.clone(),
-                        message: format!(
-                            "File is not executable (no execute permission): {display_path}"
-                        ),
-                        remediation: "Run 'chmod +x' on the file to make it executable."
-                            .to_string(),
-                        severity: HealthIssueSeverity::Error,
-                    },
-                    false,
-                ));
-            }
-            return None;
-        }
-        #[cfg(not(unix))]
-        {
-            if meta.is_file() {
-                return None;
-            }
-            return Some((
-                HealthIssue {
-                    field: field.to_string(),
-                    path: display_path.clone(),
-                    message: format!("Path exists but is not a file: {display_path}"),
-                    remediation: "Select the executable file itself.".to_string(),
-                    severity: HealthIssueSeverity::Error,
-                },
-                false,
-            ));
-        }
-    }
-
-    if path_is_executable_visible_or_host(path) {
-        return None;
-    }
-
-    if path_is_file_visible_or_host(path) {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("File is not executable (no execute permission): {display_path}"),
-                remediation: "Run 'chmod +x' on the file to make it executable.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    if path_exists_visible_or_host(path) {
-        return Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.clone(),
-                message: format!("Path exists but is not a file: {display_path}"),
-                remediation: "Select the executable file itself.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        ));
-    }
-
-    match fs::metadata(original) {
-        Ok(meta) => {
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if !meta.is_file() {
-                    return Some((
-                        HealthIssue {
-                            field: field.to_string(),
-                            path: display_path.to_string(),
-                            message: format!("Path exists but is not a file: {display_path}"),
-                            remediation: "Select the executable file itself.".to_string(),
-                            severity: HealthIssueSeverity::Error,
-                        },
-                        false,
-                    ));
-                }
-                if meta.permissions().mode() & 0o111 == 0 {
-                    return Some((
-                        HealthIssue {
-                            field: field.to_string(),
-                            path: display_path.to_string(),
-                            message: format!(
-                                "File is not executable (no execute permission): {display_path}"
-                            ),
-                            remediation: "Run 'chmod +x' on the file to make it executable."
-                                .to_string(),
-                            severity: HealthIssueSeverity::Error,
-                        },
-                        false,
-                    ));
-                }
-                None
-            }
-            #[cfg(not(unix))]
-            {
-                if !meta.is_file() {
-                    return Some((
-                        HealthIssue {
-                            field: field.to_string(),
-                            path: display_path.to_string(),
-                            message: format!("Path exists but is not a file: {display_path}"),
-                            remediation: "Select the executable file itself.".to_string(),
-                            severity: HealthIssueSeverity::Error,
-                        },
-                        false,
-                    ));
-                }
-                None
-            }
-        }
-        Err(err) if err.kind() == ErrorKind::NotFound => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Executable does not exist: {display_path}"),
-                remediation: "Re-browse to the executable or verify the path is correct."
-                    .to_string(),
-                severity: HealthIssueSeverity::Warning,
-            },
-            true,
-        )),
-        Err(err) if err.kind() == ErrorKind::PermissionDenied => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!(
-                    "Executable is not accessible (permission denied): {display_path}"
-                ),
-                remediation: "Check file permissions (e.g. chmod a+rx).".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        )),
-        Err(err) => Some((
-            HealthIssue {
-                field: field.to_string(),
-                path: display_path.to_string(),
-                message: format!("Could not access executable: {err}"),
-                remediation: "Verify the path is valid and accessible.".to_string(),
-                severity: HealthIssueSeverity::Error,
-            },
-            false,
-        )),
-    }
-}
-
-/// Check an optional path field. Empty → no issue. Missing or inaccessible → `Info`.
-fn check_optional_path(field: &str, path: &str) -> Option<HealthIssue> {
-    let display_path = display_path(path);
-    if display_path.is_empty() {
-        return None;
-    }
-
-    if path_is_file_visible_or_host(path) || path_is_dir_visible_or_host(path) {
-        return None;
-    }
-
-    if path_exists_visible_or_host(path) {
-        return None;
-    }
-
-    match fs::metadata(path.trim()) {
-        Ok(_) => None,
-        Err(err) if err.kind() == ErrorKind::NotFound => Some(HealthIssue {
-            field: field.to_string(),
-            path: display_path.to_string(),
-            message: format!("Optional path does not exist: {display_path}"),
-            remediation: format!("Browse to or clear the '{field}' field if no longer needed."),
-            severity: HealthIssueSeverity::Info,
-        }),
-        Err(_) => Some(HealthIssue {
-            field: field.to_string(),
-            path: display_path.to_string(),
-            message: format!("Optional path is not accessible: {display_path}"),
-            remediation: format!("Verify the '{field}' path or clear it if no longer needed."),
-            severity: HealthIssueSeverity::Info,
-        }),
-    }
-}
-
-/// Validates path fields of a `GameProfile` and returns a `ProfileHealthReport`.
-///
-/// Method-aware: only fields required by the resolved launch method are checked as required.
-/// All populated optional fields (icon_path, working_directory) are validated at Info severity.
-pub fn check_profile_health(name: &str, profile: &GameProfile) -> ProfileHealthReport {
-    let effective_profile = profile.effective_profile();
-    let launch_method = resolve_launch_method(&effective_profile).to_string();
-    let mut issues: Vec<HealthIssue> = Vec::new();
-    let mut has_stale = false;
-    let mut has_broken = false;
-
-    // Collect required-field results — each returns (issue, is_stale)
-    let mut required_results: Vec<Option<(HealthIssue, bool)>> = Vec::new();
-
-    // game.executable_path — required for all methods
-    required_results.push(check_required_file(
-        "game.executable_path",
-        &effective_profile.game.executable_path,
-    ));
-
-    // trainer.path — required only if non-empty
-    if !effective_profile.trainer.path.trim().is_empty() {
-        required_results.push(check_file_path(
-            "trainer.path",
-            &effective_profile.trainer.path,
-            HealthIssueSeverity::Error,
-        ));
-    }
-
-    // injection.dll_paths — each non-empty entry must exist as a file
-    for (i, dll_path) in effective_profile.injection.dll_paths.iter().enumerate() {
-        if !dll_path.trim().is_empty() {
-            required_results.push(check_file_path(
-                &format!("injection.dll_paths[{i}]"),
-                dll_path,
-                HealthIssueSeverity::Error,
-            ));
-        }
-    }
-
-    // Method-specific required fields
-    match launch_method.as_str() {
-        METHOD_STEAM_APPLAUNCH => {
-            required_results.push(check_required_directory(
-                "steam.compatdata_path",
-                &effective_profile.steam.compatdata_path,
-            ));
-            required_results.push(check_required_executable(
-                "steam.proton_path",
-                &effective_profile.steam.proton_path,
-            ));
-        }
-        METHOD_PROTON_RUN => {
-            required_results.push(check_required_directory(
-                "runtime.prefix_path",
-                &effective_profile.runtime.prefix_path,
-            ));
-            required_results.push(check_required_executable(
-                "runtime.proton_path",
-                &effective_profile.runtime.proton_path,
-            ));
-        }
-        _ => {
-            // native — no additional required path fields
-        }
-    }
-
-    // Process required-field results
-    for (issue, stale) in required_results.into_iter().flatten() {
-        if stale {
-            has_stale = true;
-        } else {
-            has_broken = true;
-        }
-        issues.push(issue);
-    }
-
-    // Optional fields — checked at Info severity regardless of method
-    if let Some(issue) = check_optional_path(
-        "steam.launcher.icon_path",
-        &effective_profile.steam.launcher.icon_path,
-    ) {
-        issues.push(issue);
-    }
-    if let Some(issue) = check_optional_path(
-        "runtime.working_directory",
-        &effective_profile.runtime.working_directory,
-    ) {
-        issues.push(issue);
-    }
-
-    // Determine overall status.
-    // Unconfigured profiles (all empty required fields) also classify as Broken per business rules;
-    // the UI presents them with badge-only (no banner) based on all issues having empty path fields.
-    let status = if has_broken {
-        HealthStatus::Broken
-    } else if has_stale {
-        HealthStatus::Stale
-    } else {
-        HealthStatus::Healthy
-    };
-
-    ProfileHealthReport {
-        name: name.to_string(),
-        status,
-        launch_method,
-        issues,
-        checked_at: Utc::now().to_rfc3339(),
-    }
-}
-
-/// Like [`batch_check_health`], but invokes `enrich` after each successful `check_profile_health`
-/// so callers can attach SQLite-backed checks (e.g. offline readiness) using one `Connection`.
-pub fn batch_check_health_with_enrich<F>(store: &ProfileStore, mut enrich: F) -> HealthCheckSummary
-where
-    F: FnMut(&str, &GameProfile, &mut ProfileHealthReport),
-{
-    let now = Utc::now().to_rfc3339();
-
-    let names = match store.list() {
-        Ok(names) => names,
-        Err(err) => {
-            return HealthCheckSummary {
-                profiles: vec![ProfileHealthReport {
-                    name: "<unknown>".to_string(),
-                    status: HealthStatus::Broken,
-                    launch_method: String::new(),
-                    issues: vec![HealthIssue {
-                        field: String::new(),
-                        path: String::new(),
-                        message: format!("Could not list profiles: {err}"),
-                        remediation: "Check filesystem permissions for the profiles directory."
-                            .to_string(),
-                        severity: HealthIssueSeverity::Error,
-                    }],
-                    checked_at: now.clone(),
-                }],
-                healthy_count: 0,
-                stale_count: 0,
-                broken_count: 1,
-                total_count: 1,
-                validated_at: now,
-            };
-        }
-    };
-
-    let mut profiles: Vec<ProfileHealthReport> = Vec::with_capacity(names.len());
-
-    for name in &names {
-        let report = match store.load(name) {
-            Ok(profile) => {
-                let mut report = check_profile_health(name, &profile);
-                enrich(name, &profile, &mut report);
-                report
-            }
-            Err(err) => ProfileHealthReport {
-                name: name.clone(),
-                status: HealthStatus::Broken,
-                launch_method: String::new(),
-                issues: vec![HealthIssue {
-                    field: String::new(),
-                    path: String::new(),
-                    message: format!("Profile could not be loaded: {err}"),
-                    remediation: "The profile TOML may be malformed. Delete and re-create the profile, or edit the file manually.".to_string(),
-                    severity: HealthIssueSeverity::Error,
-                }],
-                checked_at: Utc::now().to_rfc3339(),
-            },
-        };
-        profiles.push(report);
-    }
-
-    let healthy_count = profiles
-        .iter()
-        .filter(|r| matches!(r.status, HealthStatus::Healthy))
-        .count();
-    let stale_count = profiles
-        .iter()
-        .filter(|r| matches!(r.status, HealthStatus::Stale))
-        .count();
-    let broken_count = profiles
-        .iter()
-        .filter(|r| matches!(r.status, HealthStatus::Broken))
-        .count();
-    let total_count = profiles.len();
-
-    HealthCheckSummary {
-        profiles,
-        healthy_count,
-        stale_count,
-        broken_count,
-        total_count,
-        validated_at: now,
-    }
-}
-
-/// Runs `check_profile_health` for every profile in the store and returns a summary.
-///
-/// Errors loading individual profiles are captured as `Broken` entries and do not abort
-/// the batch — this function never propagates `ProfileStoreError` from the per-profile loop.
-pub fn batch_check_health(store: &ProfileStore) -> HealthCheckSummary {
-    batch_check_health_with_enrich(store, |_, _, _| {})
-}
-
-/// Build health issues for prefix dependency states.
-///
-/// CRITICAL: This function is pure -- takes data in, returns issues out.
-/// NO subprocess spawning, NO async, NO Command. The health scan runs at
-/// startup and must be fast and synchronous.
-pub fn build_dependency_health_issues(
-    dep_states: &[PrefixDependencyStateRow],
-    required_verbs: &[String],
-    active_prefix: &str,
-) -> Vec<HealthIssue> {
-    if required_verbs.is_empty() {
-        return Vec::new();
-    }
-
-    let mut issues = Vec::new();
-
-    // If profile has required deps but no state rows at all, deps haven't been checked
-    if dep_states.is_empty() {
-        issues.push(HealthIssue {
-            field: "prefix_dependencies".to_string(),
-            path: String::new(),
-            message: "Prefix dependencies have not been checked".to_string(),
-            remediation: "Open the profile and run a dependency check.".to_string(),
-            severity: HealthIssueSeverity::Warning,
-        });
-        return issues;
-    }
-
-    for verb in required_verbs {
-        let state = dep_states
-            .iter()
-            .find(|row| row.package_name == *verb && row.prefix_path == active_prefix)
-            .map(|row| row.state.as_str());
-
-        match state {
-            Some("installed") => {
-                // Healthy -- no issue
-            }
-            Some("missing") | Some("install_failed") => {
-                issues.push(HealthIssue {
-                    field: format!("prefix_dep:{verb}"),
-                    path: String::new(),
-                    message: format!("Required WINE dependency '{verb}' is not installed"),
-                    remediation: format!(
-                        "Install '{verb}' via the prefix dependencies panel or run winetricks manually."
-                    ),
-                    severity: HealthIssueSeverity::Warning,
-                });
-            }
-            Some("check_failed") | Some("unknown") | None => {
-                issues.push(HealthIssue {
-                    field: format!("prefix_dep:{verb}"),
-                    path: String::new(),
-                    message: format!("Dependency status unknown for '{verb}'"),
-                    remediation: "Run a dependency check to determine current status.".to_string(),
-                    severity: HealthIssueSeverity::Warning,
-                });
-            }
-            Some("user_skipped") => {
-                issues.push(HealthIssue {
-                    field: format!("prefix_dep:{verb}"),
-                    path: String::new(),
-                    message: format!("User skipped '{verb}' installation"),
-                    remediation: format!(
-                        "Install '{verb}' if the game requires it, or skip this warning."
-                    ),
-                    severity: HealthIssueSeverity::Info,
-                });
-            }
-            Some(other) => {
-                // Unknown state string -- treat as stale
-                issues.push(HealthIssue {
-                    field: format!("prefix_dep:{verb}"),
-                    path: String::new(),
-                    message: format!("Unexpected dependency state '{other}' for '{verb}'"),
-                    remediation: "Run a dependency check to refresh status.".to_string(),
-                    severity: HealthIssueSeverity::Warning,
-                });
-            }
-        }
-    }
-
-    issues
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use std::path::Path;
-    use tempfile::tempdir;
-
-    use crate::profile::{
-        GameProfile, GameSection, InjectionSection, LaunchSection, LauncherSection, RuntimeSection,
-        SteamSection, TrainerSection,
-    };
-
-    /// Create a real executable file at `path`.
-    fn make_executable(path: &Path) {
-        fs::write(path, b"#!/bin/sh\n").expect("write executable");
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let mut perms = fs::metadata(path).expect("metadata").permissions();
-            perms.set_mode(0o755);
-            fs::set_permissions(path, perms).expect("chmod");
-        }
-    }
-
-    /// Build a `GameProfile` for `steam_applaunch` where all configured paths exist.
-    fn healthy_steam_profile(tmp: &Path) -> GameProfile {
-        let game_exe = tmp.join("game.exe");
-        let trainer = tmp.join("trainer.exe");
-        let dll = tmp.join("mod.dll");
-        let compatdata = tmp.join("compatdata");
-        let proton = tmp.join("proton");
-
-        make_executable(&game_exe);
-        make_executable(&trainer);
-        fs::write(&dll, b"MZ").expect("write dll");
-        fs::create_dir_all(&compatdata).expect("mkdir compatdata");
-        make_executable(&proton);
-
-        GameProfile {
-            game: GameSection {
-                name: "Test Game".to_string(),
-                executable_path: game_exe.to_string_lossy().to_string(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection {
-                path: trainer.to_string_lossy().to_string(),
-                kind: "fling".to_string(),
-                loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-                trainer_type: "unknown".to_string(),
-                required_protontricks: Vec::new(),
-                community_trainer_sha256: String::new(),
-            },
-            injection: InjectionSection {
-                dll_paths: vec![dll.to_string_lossy().to_string()],
-                inject_on_launch: vec![true],
-            },
-            steam: SteamSection {
-                enabled: true,
-                app_id: "12345".to_string(),
-                compatdata_path: compatdata.to_string_lossy().to_string(),
-                proton_path: proton.to_string_lossy().to_string(),
-                launcher: LauncherSection {
-                    icon_path: String::new(),
-                    display_name: String::new(),
-                },
-            },
-            runtime: RuntimeSection::default(),
-            launch: LaunchSection {
-                method: "steam_applaunch".to_string(),
-                ..Default::default()
-            },
-            local_override: crate::profile::LocalOverrideSection::default(),
-        }
-    }
-
-    #[test]
-    fn healthy_profile_reports_healthy_status() {
-        let tmp = tempdir().expect("tempdir");
-        let profile = healthy_steam_profile(tmp.path());
-        let report = check_profile_health("test-game", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Healthy),
-            "expected Healthy, got {:?}; issues: {:?}",
-            report.status,
-            report.issues
-        );
-        assert!(report.issues.is_empty());
-        assert_eq!(report.name, "test-game");
-        assert_eq!(report.launch_method, "steam_applaunch");
-    }
-
-    #[test]
-    fn missing_game_exe_reports_stale() {
-        let tmp = tempdir().expect("tempdir");
-        let mut profile = healthy_steam_profile(tmp.path());
-        profile.game.executable_path = tmp
-            .path()
-            .join("nonexistent.exe")
-            .to_string_lossy()
-            .to_string();
-
-        let report = check_profile_health("stale-game", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Stale),
-            "expected Stale, got {:?}",
-            report.status
-        );
-        assert!(report
-            .issues
-            .iter()
-            .any(|i| i.field == "game.executable_path"));
-    }
-
-    #[test]
-    fn game_exe_is_directory_reports_broken() {
-        let tmp = tempdir().expect("tempdir");
-        let dir_path = tmp.path().join("itsadir");
-        fs::create_dir_all(&dir_path).expect("mkdir");
-
-        let mut profile = healthy_steam_profile(tmp.path());
-        profile.game.executable_path = dir_path.to_string_lossy().to_string();
-
-        let report = check_profile_health("broken-game", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Broken),
-            "expected Broken, got {:?}",
-            report.status
-        );
-        assert!(report
-            .issues
-            .iter()
-            .any(|i| i.field == "game.executable_path"));
-    }
-
-    #[test]
-    fn unconfigured_profile_reports_broken() {
-        let profile = GameProfile::default();
-        let report = check_profile_health("empty-profile", &profile);
-
-        // game.executable_path is required for all methods — empty → Broken
-        assert!(
-            matches!(report.status, HealthStatus::Broken),
-            "expected Broken for empty profile, got {:?}",
-            report.status
-        );
-        assert!(report
-            .issues
-            .iter()
-            .any(|i| i.field == "game.executable_path"));
-    }
-
-    #[test]
-    fn missing_proton_reports_stale_for_steam_applaunch() {
-        let tmp = tempdir().expect("tempdir");
-        let mut profile = healthy_steam_profile(tmp.path());
-        // Point proton_path at a nonexistent path
-        profile.steam.proton_path = tmp.path().join("gone_proton").to_string_lossy().to_string();
-
-        let report = check_profile_health("stale-steam", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Stale),
-            "expected Stale (missing proton), got {:?}",
-            report.status
-        );
-        assert!(report.issues.iter().any(|i| i.field == "steam.proton_path"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn proton_path_not_executable_reports_broken() {
-        let tmp = tempdir().expect("tempdir");
-        let mut profile = healthy_steam_profile(tmp.path());
-
-        // Create a non-executable file as proton
-        let non_exec = tmp.path().join("proton_no_exec");
-        fs::write(&non_exec, b"data").expect("write");
-        use std::os::unix::fs::PermissionsExt;
-        let mut perms = fs::metadata(&non_exec).expect("meta").permissions();
-        perms.set_mode(0o644);
-        fs::set_permissions(&non_exec, perms).expect("chmod");
-        profile.steam.proton_path = non_exec.to_string_lossy().to_string();
-
-        let report = check_profile_health("broken-proton", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Broken),
-            "expected Broken (non-executable proton), got {:?}",
-            report.status
-        );
-        assert!(report.issues.iter().any(|i| i.field == "steam.proton_path"));
-    }
-
-    #[test]
-    fn batch_check_health_returns_all_profiles() {
-        let tmp = tempdir().expect("tempdir");
-        let store = ProfileStore::with_base_path(tmp.path().join("profiles"));
-
-        let profile = healthy_steam_profile(tmp.path());
-        store.save("game-a", &profile).expect("save game-a");
-        store.save("game-b", &profile).expect("save game-b");
-
-        let summary = batch_check_health(&store);
-
-        assert_eq!(summary.total_count, 2);
-        assert_eq!(summary.profiles.len(), 2);
-    }
-
-    #[test]
-    fn batch_check_health_isolates_toml_parse_error() {
-        let tmp = tempdir().expect("tempdir");
-        let store = ProfileStore::with_base_path(tmp.path().join("profiles"));
-
-        // Save a valid profile
-        let profile = healthy_steam_profile(tmp.path());
-        store.save("valid-profile", &profile).expect("save valid");
-
-        // Write an intentionally malformed TOML file
-        let bad_path = store.base_path.join("broken-toml.toml");
-        fs::create_dir_all(&store.base_path).expect("mkdir profiles");
-        fs::write(&bad_path, b"[invalid toml content %%% @@").expect("write bad toml");
-
-        let summary = batch_check_health(&store);
-
-        // Should have 2 profiles total: 1 valid, 1 broken
-        assert_eq!(summary.total_count, 2);
-        assert_eq!(summary.broken_count, 1);
-
-        let broken = summary
-            .profiles
-            .iter()
-            .find(|r| r.name == "broken-toml")
-            .expect("broken-toml report missing");
-        assert!(matches!(broken.status, HealthStatus::Broken));
-        assert!(!broken.issues.is_empty());
-    }
-
-    #[test]
-    fn batch_check_health_empty_store_returns_empty_summary() {
-        let tmp = tempdir().expect("tempdir");
-        let store = ProfileStore::with_base_path(tmp.path().join("profiles"));
-
-        let summary = batch_check_health(&store);
-
-        assert_eq!(summary.total_count, 0);
-        assert_eq!(summary.healthy_count, 0);
-        assert_eq!(summary.stale_count, 0);
-        assert_eq!(summary.broken_count, 0);
-    }
-
-    #[test]
-    fn proton_run_method_checks_runtime_prefix_not_steam() {
-        let tmp = tempdir().expect("tempdir");
-
-        let game_exe = tmp.path().join("game.exe");
-        make_executable(&game_exe);
-
-        let prefix = tmp.path().join("pfx");
-        fs::create_dir_all(&prefix).expect("mkdir prefix");
-
-        let proton = tmp.path().join("proton");
-        make_executable(&proton);
-
-        let profile = GameProfile {
-            game: GameSection {
-                name: "Proton Game".to_string(),
-                executable_path: game_exe.to_string_lossy().to_string(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection::default(),
-            injection: InjectionSection::default(),
-            steam: SteamSection::default(),
-            runtime: RuntimeSection {
-                prefix_path: prefix.to_string_lossy().to_string(),
-                proton_path: proton.to_string_lossy().to_string(),
-                working_directory: String::new(),
-                steam_app_id: String::new(),
-                umu_game_id: String::new(),
-                umu_preference: None,
-            },
-            launch: LaunchSection {
-                method: "proton_run".to_string(),
-                ..Default::default()
-            },
-            local_override: crate::profile::LocalOverrideSection::default(),
-        };
-
-        let report = check_profile_health("proton-run-game", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Healthy),
-            "expected Healthy for proton_run profile with all paths present, got {:?}; issues: {:?}",
-            report.status,
-            report.issues
-        );
-    }
-
-    #[test]
-    fn host_mounted_runtime_proton_path_is_healthy() {
-        let tmp = tempdir().expect("tempdir");
-
-        let game_exe = tmp.path().join("game.exe");
-        make_executable(&game_exe);
-
-        let prefix = tmp.path().join("pfx");
-        fs::create_dir_all(&prefix).expect("mkdir prefix");
-
-        let proton = tmp.path().join("proton");
-        make_executable(&proton);
-
-        let profile = GameProfile {
-            game: GameSection {
-                name: "Proton Game".to_string(),
-                executable_path: game_exe.to_string_lossy().to_string(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection::default(),
-            injection: InjectionSection::default(),
-            steam: SteamSection::default(),
-            runtime: RuntimeSection {
-                prefix_path: format!("/run/host{}", prefix.to_string_lossy()),
-                proton_path: format!("/run/host{}", proton.to_string_lossy()),
-                working_directory: String::new(),
-                steam_app_id: String::new(),
-                umu_game_id: String::new(),
-                umu_preference: None,
-            },
-            launch: LaunchSection {
-                method: "proton_run".to_string(),
-                ..Default::default()
-            },
-            local_override: crate::profile::LocalOverrideSection::default(),
-        };
-
-        let report = check_profile_health("proton-run-host-mounted", &profile);
-
-        assert!(
-            matches!(report.status, HealthStatus::Healthy),
-            "expected Healthy for host-mounted proton_run profile, got {:?}; issues: {:?}",
-            report.status,
-            report.issues
-        );
-    }
-
-    // Prefix dependency health issue tests
-
-    fn make_dep_row(package: &str, state: &str) -> PrefixDependencyStateRow {
-        PrefixDependencyStateRow {
-            id: 0,
-            profile_id: "test".to_string(),
-            package_name: package.to_string(),
-            prefix_path: "/tmp/pfx".to_string(),
-            state: state.to_string(),
-            checked_at: Some("2026-01-01T00:00:00Z".to_string()),
-            installed_at: None,
-            last_error: None,
-            created_at: "2026-01-01T00:00:00Z".to_string(),
-            updated_at: "2026-01-01T00:00:00Z".to_string(),
-        }
-    }
-
-    #[test]
-    fn installed_verbs_produce_no_issues() {
-        let deps = vec![
-            make_dep_row("vcrun2019", "installed"),
-            make_dep_row("dotnet48", "installed"),
-        ];
-        let required = vec!["vcrun2019".to_string(), "dotnet48".to_string()];
-        let issues = build_dependency_health_issues(&deps, &required, "/tmp/pfx");
-        assert!(issues.is_empty());
-    }
-
-    #[test]
-    fn missing_verb_produces_warning() {
-        let deps = vec![make_dep_row("vcrun2019", "missing")];
-        let required = vec!["vcrun2019".to_string()];
-        let issues = build_dependency_health_issues(&deps, &required, "/tmp/pfx");
-        assert_eq!(issues.len(), 1);
-        assert!(matches!(issues[0].severity, HealthIssueSeverity::Warning));
-        assert!(issues[0].message.contains("not installed"));
-    }
-
-    #[test]
-    fn unknown_state_produces_warning() {
-        let deps = vec![];
-        let required = vec!["vcrun2019".to_string()];
-        let issues = build_dependency_health_issues(&deps, &required, "/tmp/pfx");
-        // Empty dep_states with required verbs → "not been checked" warning
-        assert_eq!(issues.len(), 1);
-        assert!(issues[0].message.contains("not been checked"));
-    }
-
-    #[test]
-    fn skipped_verb_produces_info() {
-        let deps = vec![make_dep_row("vcrun2019", "user_skipped")];
-        let required = vec!["vcrun2019".to_string()];
-        let issues = build_dependency_health_issues(&deps, &required, "/tmp/pfx");
-        assert_eq!(issues.len(), 1);
-        assert!(matches!(issues[0].severity, HealthIssueSeverity::Info));
-        assert!(issues[0].message.contains("skipped"));
-    }
-
-    #[test]
-    fn empty_required_verbs_produce_no_issues() {
-        let deps = vec![make_dep_row("vcrun2019", "installed")];
-        let required: Vec<String> = vec![];
-        let issues = build_dependency_health_issues(&deps, &required, "/tmp/pfx");
-        assert!(issues.is_empty());
-    }
-}
