+++ src/crosshook-native/crates/crosshook-core/src/profile/health/mod.rs
@@ -1,0 +1,12 @@
+mod checks;
+mod dependency;
+mod path_probe;
+mod profile;
+mod types;
+
+pub use dependency::*;
+pub use profile::*;
+pub use types::*;
+
+#[cfg(test)]
+mod tests;
