+++ src/crosshook-native/crates/crosshook-core/src/profile/health/tests/mod.rs
@@ -1,0 +1,3 @@
+pub(super) mod dependency;
+pub(super) mod fixtures;
+pub(super) mod profile;
